final changeset to implement different internal/external infilling specifications (where external infillings are defined as the part of the slice that is not covered in both Z sides, above and below, by other slices close in Z). This is the final piece of the puzzle that started with the commit that refactored the monolithic Multislicer::applyProcess() in two phases

+++ interfaces/slicermanager.cpp
@@ -271,7 +271,7 @@ bool RawSlicerManager::start(const char * fname) {
         return false;
     }
     
-    numRecords = fileheader.numRecords;
+    numRecords = (int)fileheader.numRecords;
 
     iop_f = IOPaths(f);
